Move code to HOC
